Elimina la vista principal `MainView` y refactoriza las vistas de inicio y perfil para utilizar `HomeView` y `ProfileView` respectivamente. Se implementa un nuevo diseño de navegación que utiliza `NavBar` y se ajustan las rutas en la configuración de navegación para reflejar estos cambios. Además, se simplifica la lógica de estado en las vistas de perfil y de inicio.
