+++ lib/presentation/profile/profile_view.dart
@@ -1,270 +1,12 @@
 import 'package:flutter/material.dart';
 //widgets
-import '../../widgets/master/bg.dart';
-import '../../widgets/glass/glass_container.dart';
-import '../home/nav_bar.dart';
+import 'profile_content.dart';
 
-class ProfileView extends StatefulWidget {
+class ProfileView extends StatelessWidget {
   const ProfileView({super.key});
 
   @override
-  State<ProfileView> createState() => _ProfileViewState();
-}
-
-class _ProfileViewState extends State<ProfileView> {
-  bool _darkMode = true;
-  bool _notifications = true;
-  bool _saveHistory = true;
-  bool _privacyMode = true;
-
-  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: Bg.decoration,
-        child: SafeArea(
-          child: Column(
-            children: [
-              Expanded(
-                child: SingleChildScrollView(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 32.0,
-                    vertical: 20.0,
-                  ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // Título "Cuenta"
-                      _buildSectionTitle("Cuenta"),
-
-                      const SizedBox(height: 16),
-
-                      // Mi Perfil
-                      _buildProfileItem(
-                        icon: Icons.person_outline,
-                        title: "Mi Perfil",
-                        subtitle: "Administra tu información Personal",
-                        onTap: () {
-                          // Navegación a editar perfil
-                        },
-                      ),
-
-                      const SizedBox(height: 32),
-
-                      // Título "Preferencias"
-                      _buildSectionTitle("Preferencias"),
-
-                      const SizedBox(height: 16),
-
-                      // Modo Oscuro
-                      _buildToggleItem(
-                        icon: Icons.dark_mode_outlined,
-                        title: "Modo Oscuro",
-                        subtitle: "Cambia entre el modo claro y oscuro",
-                        value: _darkMode,
-                        onChanged: (value) {
-                          setState(() {
-                            _darkMode = value;
-                          });
-                        },
-                      ),
-
-                      const SizedBox(height: 12),
-
-                      // Notificaciones
-                      _buildToggleItem(
-                        icon: Icons.notifications_outlined,
-                        title: "Notificaciones",
-                        subtitle: "Recibe alertas y recordatorios",
-                        value: _notifications,
-                        onChanged: (value) {
-                          setState(() {
-                            _notifications = value;
-                          });
-                        },
-                      ),
-
-                      const SizedBox(height: 12),
-
-                      // Guardar el Historial
-                      _buildToggleItem(
-                        icon: Icons.history,
-                        title: "Guardar el Historial",
-                        subtitle: "Conserva tus análisis anteriores",
-                        value: _saveHistory,
-                        onChanged: (value) {
-                          setState(() {
-                            _saveHistory = value;
-                          });
-                        },
-                      ),
-
-                      const SizedBox(height: 32),
-
-                      // Título "Políticas de Privacidad"
-                      _buildSectionTitle("Políticas de Privacidad"),
-
-                      const SizedBox(height: 16),
-
-                      // Privacidad
-                      _buildToggleItem(
-                        icon: Icons.privacy_tip_outlined,
-                        title: "Modo Privado",
-                        subtitle: "Cambia entre el modo claro y oscuro",
-                        value: _privacyMode,
-                        onChanged: (value) {
-                          setState(() {
-                            _privacyMode = value;
-                          });
-                        },
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-
-              // Navigation Bar
-              NavBar.buildNavBar(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSectionTitle(String title) {
-    return Text(
-      title,
-      style: const TextStyle(
-        color: Colors.white,
-        fontSize: 24,
-        fontWeight: FontWeight.w600,
-      ),
-    );
-  }
-
-  Widget _buildProfileItem({
-    required IconData icon,
-    required String title,
-    required String subtitle,
-    required VoidCallback onTap,
-  }) {
-    return GestureDetector(
-      onTap: onTap,
-      child: GlassContainer(
-        borderRadius: 20,
-        padding: const EdgeInsets.all(16),
-        child: Row(
-          children: [
-            // Icono
-            Container(
-              width: 50,
-              height: 50,
-              decoration: BoxDecoration(
-                color: Colors.white.withValues(alpha: 0.1),
-                borderRadius: BorderRadius.circular(12),
-              ),
-              child: Icon(icon, color: Colors.white, size: 24),
-            ),
-
-            const SizedBox(width: 16),
-
-            // Textos
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    title,
-                    style: const TextStyle(
-                      color: Colors.white,
-                      fontSize: 16,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(
-                    subtitle,
-                    style: TextStyle(
-                      color: Colors.white.withValues(alpha: 0.7),
-                      fontSize: 14,
-                      fontWeight: FontWeight.w300,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildToggleItem({
-    required IconData icon,
-    required String title,
-    required String subtitle,
-    required bool value,
-    required ValueChanged<bool> onChanged,
-  }) {
-    return GlassContainer(
-      borderRadius: 20,
-      padding: const EdgeInsets.all(16),
-      child: Row(
-        children: [
-          // Icono
-          Container(
-            width: 50,
-            height: 50,
-            decoration: BoxDecoration(
-              color: Colors.white.withValues(alpha: 0.1),
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Icon(icon, color: Colors.white, size: 24),
-          ),
-
-          const SizedBox(width: 16),
-
-          // Textos
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  title,
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontSize: 16,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-                const SizedBox(height: 4),
-                Text(
-                  subtitle,
-                  style: TextStyle(
-                    color: Colors.white.withValues(alpha: 0.7),
-                    fontSize: 14,
-                    fontWeight: FontWeight.w300,
-                  ),
-                ),
-              ],
-            ),
-          ),
-
-          // Toggle Switch
-          Transform.scale(
-            scale: 0.8,
-            child: Switch(
-              value: value,
-              onChanged: onChanged,
-              activeColor: Colors.white,
-              activeTrackColor: const Color(0xFF6366F1),
-              inactiveThumbColor: Colors.white.withValues(alpha: 0.7),
-              inactiveTrackColor: Colors.white.withValues(alpha: 0.2),
-            ),
-          ),
-        ],
-      ),
-    );
+    return const ProfileContent();
   }
 }
