+++ lib/routes/routes.dart
@@ -1,21 +1,16 @@
 import 'package:get/get.dart';
 
-//navigation
-import '../presentation/home/nav_bar.dart';
-
 //views
 import '../presentation/views/welcome_view.dart';
-import '../presentation/main_view.dart';
+import '../presentation/main_layout.dart';
 
 //routes
 class Routes {
   static final List<GetPage> pages = [
-    //nav bar
-    GetPage(name: '/nav_bar', page: () => NavBar.buildNavBar()),
     //welcome
     GetPage(name: '/', page: () => const WelcomeView()),
-    //main app (home + profile con nav fijo)
-    GetPage(name: '/home', page: () => const MainView()),
-    GetPage(name: '/profile', page: () => const MainView()),
+    //main app (maneja home y profile internamente)
+    GetPage(name: '/home', page: () => const MainLayout(initialIndex: 0)),
+    GetPage(name: '/profile', page: () => const MainLayout(initialIndex: 1)),
   ];
 }
